Extract Mesh class to separate file
