+++ mikeio/dfsu/__init__.py
@@ -1,9 +1,5 @@
-from ._dfsu import Mesh, _write_dfsu
+from ._dfsu import _write_dfsu, Dfsu2DH
+from ._mesh import Mesh
 from ._factory import Dfsu
 
-# Alias for fmskill/modelskill
-from ._dfsu import Dfsu2DH
-
-_Dfsu = Dfsu2DH
-
-__all__ = ["Mesh", "Dfsu", "_write_dfsu", "Dfsu2DH", "_Dfsu"]
+__all__ = ["Mesh", "Dfsu", "_write_dfsu", "Dfsu2DH"]
